Add utility for changing sandbox mango hooks
